create 'asis' vignettes
Prior to this change R CMD check would fail to rebuild the vignette.
The primary reason was lack of access to my private key to access the
vault, or actucally, access to the vault since the vault is not part of
the source package.  I would expect, even if the vault was accessable
that the external curl requests would cause an issue.  As such, the
vignette needs to be built by the developer and the package needs to
have a static vignette.