create S3 method for `format_record`
Prior to this change `format_record` required the user to pass in both
the record and metadata, or column types, which was cumbersome and not
needed since a `rcer_rccore` object exists.

This change creates a S3 method for `rcer_rcrore` and the previous
method is now the `rcer_raw_record` and `rcer_record` method.

This change also introduces the warning message noted in issue #17